Move service to modules folder
