+++ packages/client-web/src/components/FeedLogo/index.tsx
@@ -1,4 +1,4 @@
-import { FeedName } from "../../services/feeds/types";
+import { FeedName } from "../../modules/feeds/types";
 
 const logoMap = {
   [FeedName.Toshl]: {
