+++ packages/client-web/src/components/FeedSettings/index.tsx
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
 import { ButtonPlain, HorizontalLoader } from "@quoll/ui-components";
-import { FeedName } from "../../services/feeds/types";
+import { FeedName } from "../../modules/feeds/types";
 import { FeedState } from "../../modules/feeds/model/store";
 import FeedLogo from "../FeedLogo";
